@@ -1,9 +1,13 @@
-package com.prx.security;
+package com.prx.security.service;
 
+import com.prx.security.exception.CertificateSecurityException;
 import io.jsonwebtoken.Claims;
+import io.jsonwebtoken.ExpiredJwtException;
 import org.apache.commons.lang.NotImplementedException;
 
-import java.util.Date;
+import java.time.LocalDateTime;
+import java.time.ZoneId;
+import java.util.Map;
 import java.util.Optional;
 
 import static com.prx.security.constant.ConstantApp.SESSION_TOKEN_KEY;
@@ -22,12 +26,15 @@ public interface SessionJwtService {
      * @param token the JWT token
      * @return true if the token is expired, false otherwise
      */
-    default boolean isTokenExpired(String token) {
+    default boolean isTokenExpired(String token) throws Exception {
         try {
             Claims claims = getTokenClaims(token);
-            return claims.getExpiration().before(new Date());
-        } catch (Exception e) {
+            return LocalDateTime.now().isAfter(claims.getExpiration().toInstant().atZone(ZoneId.systemDefault()).toLocalDateTime());
+//            return claims.getExpiration().before(new Date());
+        } catch (ExpiredJwtException e) {
             return true;
+        } catch (Exception e) {
+            throw e;
         }
     }
 
@@ -40,8 +47,10 @@ public interface SessionJwtService {
     default boolean isValid(String token) {
         try {
             Claims claims = getTokenClaims(token);
-            return SESSION_TOKEN_KEY.equals(getTokenClaims(token).get("type")) && new Date().before(claims.getExpiration());
-        } catch (Exception e) {
+            return SESSION_TOKEN_KEY.equals(claims.get("type")) && !isTokenExpired(token);
+        } catch (ExpiredJwtException | NotImplementedException e) {
+            return false;
+        }  catch (Exception e) {
             return false;
         }
     }
@@ -62,12 +71,13 @@ public interface SessionJwtService {
     }
 
     /**
-     * Generates a session token for the given username.
+     * Generates a session token for the given username and parameters.
      *
-     * @param username the username for which to generate the session token
+     * @param username   the username for which the session token is generated
+     * @param parameters a map of additional parameters to include in the token
      * @return the generated session token
      */
-    String generateSessionToken(String username);
+    String generateSessionToken(String username, Map<String, String> parameters);
 
     /**
      * Retrieves the claims from the given token.
@@ -75,7 +85,7 @@ public interface SessionJwtService {
      * @param token the JWT token
      * @return the claims contained in the token
      */
-    default Claims getTokenClaims(String token) {
+    default Claims getTokenClaims(String token) throws CertificateSecurityException {
         throw new NotImplementedException();
     }
 
